refactor: Doing some housekeeping refactoring
* Remove the logger from function params
  * A more modern approach is to use the logger in the context directly
* More succinct usage of err with an `if err := fn(); err != nil`
