+++ internal/reconcilers/service.go
@@ -7,30 +7,29 @@ import (
 
 	ctrl "sigs.k8s.io/controller-runtime"
 
-	"github.com/go-logr/logr"
 	"github.io/opdev/docling-operator/api/v1alpha1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/util/intstr"
 	"sigs.k8s.io/controller-runtime/pkg/client"
+	logf "sigs.k8s.io/controller-runtime/pkg/log"
 )
 
 type ServiceReconciler struct {
 	client.Client
-	Log    logr.Logger
 	Scheme *runtime.Scheme
 }
 
-func NewServiceReconciler(client client.Client, log logr.Logger, scheme *runtime.Scheme) *ServiceReconciler {
+func NewServiceReconciler(client client.Client, scheme *runtime.Scheme) *ServiceReconciler {
 	return &ServiceReconciler{
 		Client: client,
-		Log:    log,
 		Scheme: scheme,
 	}
 }
 
 func (r *ServiceReconciler) Reconcile(ctx context.Context, doclingServ *v1alpha1.DoclingServ) (bool, error) {
+	log := logf.FromContext(ctx)
 	service := &corev1.Service{ObjectMeta: metav1.ObjectMeta{Name: doclingServ.Name + "-service", Namespace: doclingServ.Namespace}}
 	_, err := controllerutil.CreateOrUpdate(ctx, r.Client, service, func() error {
 		labels := labelsForDocling(doclingServ.Name)
@@ -46,12 +45,11 @@ func (r *ServiceReconciler) Reconcile(ctx context.Context, doclingServ *v1alpha1
 		_ = ctrl.SetControllerReference(doclingServ, service, r.Scheme)
 		return nil
 	})
-
 	if err != nil {
-		r.Log.Error(err, "Error reconciling Service", "Service.Namespace", service.Namespace, "Service.Name", service.Name)
+		log.Error(err, "Error reconciling Service", "Service.Namespace", service.Namespace, "Service.Name", service.Name)
 		return true, err
 	}
 
-	r.Log.Info("Successfully reconciled Service", "Service.Namespace", service.Namespace, "Service.Name", service.Name)
+	log.Info("Successfully reconciled Service", "Service.Namespace", service.Namespace, "Service.Name", service.Name)
 	return false, nil
 }
